pyproject.toml
[project]
name = "differ"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "jiwer>=4.0.0",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
